Update EditDistance_DynamicProgramming.py
Edit Distance (Dynamic Programming)
===================================
The edit distance between two strings is the minimum number of operations (insertions, deletions, and substitutions of symbols) to transform one string into another. It is a measure of similarity of two strings. The goal is to implement the algorithm for computing the edit distance between two strings.
Input Format: Each of the two lines of the input contains a string consisting of lower case latin letters.
Constraints: The length of both strings is at least 1 and at most 100.
Output Format: Output the edit distance between the given two strings.